+++ src/setup.jl
@@ -32,6 +32,7 @@ export coordinate_from_string
 export coriolis
 export CT
 export depth_from_pressure
+export despike
 export fix_gsw_bad_code
 export fix_gsw_bad_code!
 export geod_distance
@@ -73,6 +74,8 @@ export read_echosounder
 export read_section
 export read_topography
 export rename_data
+export running_mean
+export running_median
 export salinity_from_conductivity
 export SA
 export scale_bar